// src/app/modules/configuration/finance/models/invoice-adjustment-type-status-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Model to update invoice adjustment type status
 */
export interface InvoiceAdjustmentTypeStatusView { 
    /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or Sets IsInActive
     */
    isInActive: number;
}