// src/app/modules/configuration/finance/models/invoice-adjustment-type-update-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Model for updating invoice adjustment type
 */
export interface InvoiceAdjustmentTypeUpdateView { 
    /**
     * Gets or Sets Id
     */
    id: string;
    /**
     * Gets or Sets Name
     */
    name: string;
    /**
     * Gets or Sets IsPositiveAdjustment
     */
    isPositiveAdjustment: number;
}